requirements.txt
aioredis==1.3.1
amqp==5.1.1
api-client==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
billiard==4.1.0
celery==5.3.1
certifi==2022.5.18.1
cffi==1.15.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
crispy-bootstrap5==0.7
cryptography==37.0.2
daphne==4.0.0
Deprecated==1.2.13
Django==4.2.3
django-auth-ldap==4.3.0
django-cors-headers==3.13.0
django-crispy-forms==2.0
django-filter==23.2
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
kombu==5.3.1
mistune==3.0.1
msgpack==1.0.4
packaging==21.3
phonenumberslite==8.12.56
prompt-toolkit==3.0.29
psycopg2==2.9.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.15.1
PyJWT==2.8.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-ldap==3.4.0
pytz==2022.1
redis==4.6.0
requests==2.31.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tenacity==8.2.2
termcolor==2.3.0
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2023.3
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zope.interface==5.4.0
